Add opt-in stream-based API to buffer reads
There is significant overhead calling the stream reading functions from
nibbles. There is even more overhead when gray streams (and their associated
generic function dispatch) are involved.

However, the existing approach is very safe. It will never try to read beyond
the end of a compressed data stream. If it did, it may consume bytes the user
did not intend or may even block forever, waiting for an EOF.

Address this by adding the :BUFFER-SIZE option to MAKE-OCTET-STREAM-CONTEXT. If
provided, we buffer reads from the stream by this amount. However, the caller
is responsible for providing an accurate :END, beyond which 3BZ will never
read.